Praise linter
