+++ src/main/java/zd/zero/waifu/motivator/plugin/personality/IdlePersonalityCore.kt
@@ -13,8 +13,8 @@ class IdlePersonalityCore : PersonalityCore {
     private var isEventDisplayed = false
 
     override fun processMotivationEvent(
-            motivationEvent: MotivationEvent,
-            emotionalState: EmotionalState
+        motivationEvent: MotivationEvent,
+        emotionalState: EmotionalState
     ) {
         if (isEventDisplayed.not()) {
             val project = motivationEvent.project
